+++ Blazor.Wizard.Demo.Tests/AgeRuleTests.cs
@@ -1,4 +1,4 @@
-using Blazor.Wizard.Demo.Models;
+using Blazor.Wizard.Demo.Models.Person;
 using FluentAssertions;
 using Xunit;
