+++ Blazor.Wizard.Demo.Tests/PersonModelResultBuilderTests.cs
@@ -1,7 +1,7 @@
 using Blazor.Wizard.Core;
 using Blazor.Wizard.Demo.Components.WizardLogic;
 using Blazor.Wizard.Demo.Components.WizardLogic.Person;
-using Blazor.Wizard.Demo.Models;
+using Blazor.Wizard.Demo.Models.Person;
 using FluentAssertions;
 using Xunit;
