+++ Blazor.Wizard.Demo/Components/Person/PensionInfoForm.razor
@@ -1,4 +1,5 @@
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Person
 <EditForm EditContext="@EditContext">
     <div class="mb-3">
         <h4>Pension Information</h4>
