+++ Blazor.Wizard.Demo/Components/Person/PersonInfoForm.razor
@@ -1,4 +1,5 @@
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Person
 <EditForm EditContext="@EditContext">
     <DataAnnotationsValidator />
     <div class="mb-3">
