+++ Blazor.Wizard.Demo/Components/Person/PersonWizardDialog.razor
@@ -1,5 +1,6 @@
 @using Blazor.Wizard.Demo.Components.Common
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Person
 <WizardDialog TResult="PersonModel"
               Title="Person Information Wizard"
               ViewModel="@_viewModel"
