+++ Blazor.Wizard.Demo/Components/Person/SummaryView.razor
@@ -1,4 +1,5 @@
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Person
 
 <h5 class="mb-3">Review Your Information</h5>
