+++ Blazor.Wizard.Demo/Components/Questionary/QuestionaryReportStep.razor
@@ -1,4 +1,5 @@
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Questionary
 <div>
     <h4>Summary Report</h4>
     <ul>
