+++ Blazor.Wizard.Demo/Components/Questionary/QuestionaryStep3.razor
@@ -1,4 +1,5 @@
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Questionary
 <EditForm EditContext="EditContext">
   <DataAnnotationsValidator/>
   <div>
