+++ Blazor.Wizard.Demo/Components/Questionary/QuestionaryWizardDialog.razor
@@ -1,5 +1,6 @@
 @using Blazor.Wizard.Demo.Components.Common
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Questionary
 @implements IDisposable
 
 <WizardDialog TResult="QuestionaryModel"
