+++ Blazor.Wizard.Demo/Components/WizardLogic/Person/AddressStepLogic.cs
@@ -1,5 +1,5 @@
 using Blazor.Wizard.Core;
-using Blazor.Wizard.Demo.Models;
+using Blazor.Wizard.Demo.Models.Person;
 using Blazor.Wizard.Interfaces;
 
 namespace Blazor.Wizard.Demo.Components.WizardLogic.Person;
