+++ Blazor.Wizard.Demo/Components/WizardLogic/Person/PersonModelMapper.cs
@@ -1,4 +1,4 @@
-using Blazor.Wizard.Demo.Models;
+using Blazor.Wizard.Demo.Models.Person;
 using Blazor.Wizard.Interfaces;
 
 namespace Blazor.Wizard.Demo.Components.WizardLogic.Person;
