+++ Blazor.Wizard.Demo/Components/WizardLogic/Person/PersonModelResultBuilder.cs
@@ -1,4 +1,4 @@
-using Blazor.Wizard.Demo.Models;
+using Blazor.Wizard.Demo.Models.Person;
 using Blazor.Wizard.Interfaces;
 using Blazor.Wizard.Obsolete;
