+++ Blazor.Wizard.Demo/Components/WizardLogic/Questionary/QuestionaryStepRegistry.cs
@@ -1,6 +1,6 @@
 using Blazor.Wizard.Core;
 using Blazor.Wizard.Demo.Components.Questionary;
-using Blazor.Wizard.Demo.Models;
+using Blazor.Wizard.Demo.Models.Questionary;
 using Blazor.Wizard.Interfaces;
 
 namespace Blazor.Wizard.Demo.Components.WizardLogic.Questionary;
