+++ Blazor.Wizard.Demo/Pages/WizardDemo.razor
@@ -2,6 +2,8 @@
 @using Blazor.Wizard.Demo.Components.Person
 @using Blazor.Wizard.Demo.Components.Questionary
 @using Blazor.Wizard.Demo.Models
+@using Blazor.Wizard.Demo.Models.Person
+@using Blazor.Wizard.Demo.Models.Questionary
 
 <h2>Wizard Demo - Native Dialog</h2>
